app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.0.1
torch==2.9.1
torchvision==0.24.1
matplotlib==3.9.2
